requirements_with_github.txt
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
PyGithub>=1.55
python-dateutil>=2.8.0
click>=8.0.0
requests>=2.26.0